Comentários
comnetários funcionais, apangando a <div> que diz que não existem comentários quando se cria, possível apagar comentários, e volta  a aparecer que não tem comentários caso sejam apagados
